+++ api/typescript/src/index.ts
@@ -1,0 +1,4 @@
+//Example API server RPC API Bindings
+export { ExampleAPIServerServiceClient } from "./example_api_server_rpc_api_bindings/example_api_server_grpc_pb";
+export { AddPersonArgs, GetPersonArgs, GetPersonResponse, IncrementBooksReadArgs} from "./example_api_server_rpc_api_bindings/example_api_server_pb"
+