+++ docs/changelog.md
@@ -1,2 +1,6 @@
 # TBD
-* Initial commit
+### Features
+* Created example API server protobuf service definition in `example_api_server.proto` file
+* Added `regenerate-protobuf-output.sh` shell script to regenerate rpc Golang and Typescript binding files
+* Updated `index.ts` file to export Typescript binding class
+* Configured `package.json` file with the repository information